fix: separated ui and backend constants
