+++ src/pages/ExamsPage/components/ResetExamAttemptButton.jsx
@@ -15,6 +15,7 @@ const ResetExamAttemptButton = ({ username, examName, attemptId }) => {
   return (
     <>
       <div className="d-flex">
+        {messages.ResetExamAttemptButtonTitle.defaultMessage};
         <Button variant="link" size="sm" onClick={open}>
           {/* TODO: Figure out why this has an extra semicolon on it by default */}
           {formatMessage(messages.ResetExamAttemptButtonTitle)};
